Feat(elemnt): pagina de administracion de puestos de trabajo.  implmentacion Toast, alert y dropdown

+++ reservas-asw-front/src/app/admin/pages/workstation/list-workstation/list-workstation.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { AlertsService } from 'src/app/services/alerts.service';
+import { ToastsService } from 'src/app/services/toasts.service';
 import { RouteName } from '../../../../../utils/enums';
 import { Workstation, WorkstationResponse, WorkstationsResponse } from '../../../interfaces/workstation.interfaces';
 import { WorkstationsService } from '../../../services/workstations.service';
@@ -14,7 +16,9 @@ export class ListWorkstationComponent implements OnInit {
   workstations: Workstation[] = [];
 
   constructor(
-    private workstationsService: WorkstationsService
+    private workstationsService: WorkstationsService,
+    private toastService: ToastsService,
+    private alertsService: AlertsService
   ) { }
 
   ngOnInit(): void {
@@ -28,10 +32,30 @@ export class ListWorkstationComponent implements OnInit {
       );
   }
   deleteWorkstation(id: number): void {
-    this.workstationsService.deleteWorkstation(id)
-      .subscribe(
-        (workstationResponse: WorkstationResponse) => this.workstations = this.workstations.filter((workstation: Workstation) => workstation.idPuestoTrabajo !== id)
-      );
+    this.alertsService.showConfirmDialog({
+      message: '¿Desea eliminar el puesto de trabajo, esta acción no se podrá revertir?',
+      header: 'Eliminar puesto de trabajo',
+    })
+    .then(resp =>{
+      if(resp){
+        this.workstationsService.deleteWorkstation(id)
+        .subscribe(
+          (workstationResponse: WorkstationResponse) => {
+            this.workstations = this.workstations.filter((workstation: Workstation) => workstation.idPuestoTrabajo !== id);
+            this.toastService.showToastSuccess({ summary: 'Dominio eliminado', detail: 'El dominio ha sido eliminado correctamente.' });
+          }
+        );
+
+      } else {
+        return;
+      }
+    })
+    .catch(console.log);
+
+
+
+    
+      
   }
 
 }
